Share AArch64 gate recovery
