+++ litebox_platform_linux_userland/Cargo.toml
@@ -8,7 +8,7 @@ arrayvec = { version = "0.7.6", default-features = false }
 cfg-if = "1.0.0"
 libc = { version = "0.2.169", default-features = false }
 litebox = { path = "../litebox/", version = "0.1.0" }
-litebox_common_linux = { path = "../litebox_common_linux", version = "0.1.0" }
+litebox_common_linux = { path = "../litebox_common_linux", version = "0.1.0", features = ["aarch64_gate_recovery"] }
 litebox_common_optee = { path = "../litebox_common_optee", version = "0.1.0", default-features = false, optional = true }
 litebox_platform = { path = "../litebox_platform", version = "0.1.0" }
 spin = "0.9.8"
