+++ litebox_platform_linux_userland/src/aarch64.rs
@@ -5,19 +5,9 @@
 //! host anchor; the guest thread pointer is virtualized in memory.
 
 use super::*;
+use litebox_common_linux::gate_recovery::GateRuntimeState;
+pub(super) use litebox_common_linux::gate_recovery::{Aarch64GateSignalResult, GateInterruption};
 use litebox_common_linux::signal::aarch64::GuestVectorState;
-use litebox_syscall_rewriter::aarch64::{
-    GATE_ALIGNMENT, GATE_PC_CANDIDATE_COUNT, GATE_SLOT_SIZES, GateMetadata, MSR_FRAME_BYTES,
-    MSR_FRAME_OFF_VALUE, MrsTpidrGateOffset, MrsTpidrValueSource, MsrTpidrFrameState,
-    MsrTpidrGateOffset, RuntimeAccess, SVC_FRAME_BYTES, SVC_GATE_BYTES, SVC_SLOT_BYTES,
-    SvcFrameState, SvcGateOffset,
-};
-#[cfg(feature = "aarch64_virtualize_x18")]
-use litebox_syscall_rewriter::aarch64::{
-    X18_FRAME_BYTES, X18AdrOffset, X18CompareBranchOffset, X18FaultAttribution, X18FrameState,
-    X18GateOffset, X18Resume, X18StackWritebackFrameState, X18StackWritebackOffset,
-    X18StackWritebackSpSource, X18ValueSource, x18_branch_recovery_plan,
-};
 
 /// Overwrites the destination with the TLS control-block address derived from
 /// the host anchor in `TPIDR_EL0`. Callbacks entered directly from a rewritten
@@ -1353,112 +1343,6 @@ pub(super) fn sync_instruction_stream(range: core::ops::Range<usize>) {
     }
 }
 
-// The canonical context must remain stack-backed: boxing here would allocate in
-// a signal handler, violating the async-signal-safe contract.
-#[allow(clippy::large_enum_variant)]
-pub(super) enum Aarch64GateSignalResult {
-    NotGate,
-    Canonicalized(litebox_common_linux::PtRegs),
-    #[cfg_attr(
-        not(feature = "aarch64_virtualize_x18"),
-        expect(
-            dead_code,
-            reason = "only the x18 indirect-branch gate resumes the guest transparently"
-        )
-    )]
-    ResumeGuest(litebox_common_linux::PtRegs),
-    PreserveSavedContext,
-    InvalidRuntimeState,
-}
-
-#[derive(Clone, Copy, PartialEq, Eq)]
-pub(super) enum GateInterruption {
-    /// A hardware fault attributed to the current gate instruction.
-    Synchronous,
-    /// A synchronous breakpoint caused by the current gate instruction.
-    Breakpoint,
-    /// An independently delivered signal interrupting the gate.
-    Asynchronous,
-}
-
-#[derive(Clone, Copy)]
-struct GateRuntimeState {
-    guest_thread_pointer_addr: usize,
-    expected_outbound_stub: usize,
-    expected_outbound_pc: usize,
-}
-
-fn read_usize(read: &mut impl FnMut(usize, &mut [u8]) -> bool, address: usize) -> Option<usize> {
-    let mut bytes = [0u8; size_of::<usize>()];
-    read(address, &mut bytes).then(|| usize::from_ne_bytes(bytes))
-}
-
-#[cfg(feature = "aarch64_virtualize_x18")]
-fn recover_x18_frame(
-    state: X18FrameState,
-    signal_sp: usize,
-    read: &mut impl FnMut(usize, &mut [u8]) -> bool,
-) -> Option<(usize, Option<[usize; 2]>)> {
-    let (frame_sp, restore_registers, pop_frame) = match state {
-        X18FrameState::Absent => return Some((signal_sp, None)),
-        X18FrameState::AtSpRegistersLive => (signal_sp, false, true),
-        X18FrameState::AtSpRestoreRegisters => (signal_sp, true, true),
-        X18FrameState::BelowSpRestoreRegisters => {
-            let frame_sp = signal_sp.wrapping_sub(usize::from(X18_FRAME_BYTES));
-            (frame_sp, true, false)
-        }
-    };
-    let restored = if restore_registers {
-        let mut frame = [[0u8; size_of::<usize>()]; 2];
-        if !read(frame_sp, frame.as_flattened_mut()) {
-            return None;
-        }
-        Some(frame.map(usize::from_ne_bytes))
-    } else {
-        None
-    };
-    let final_sp = if pop_frame {
-        frame_sp.wrapping_add(usize::from(X18_FRAME_BYTES))
-    } else {
-        signal_sp
-    };
-    Some((final_sp, restored))
-}
-
-#[cfg(feature = "aarch64_virtualize_x18")]
-fn apply_x18_recovery(
-    canonical: &mut litebox_common_linux::PtRegs,
-    scratch: u8,
-    anchor_scratch: u8,
-    restored_registers: Option<[usize; 2]>,
-    guest_sp: usize,
-    guest_x18: usize,
-    resume_pc: usize,
-) {
-    if let Some([saved_scratch, saved_anchor]) = restored_registers {
-        canonical.regs[usize::from(scratch)] = saved_scratch;
-        canonical.regs[usize::from(anchor_scratch)] = saved_anchor;
-    }
-    canonical.sp = guest_sp;
-    canonical.regs[18] = guest_x18;
-    canonical.pc = resume_pc;
-}
-
-#[cfg(feature = "aarch64_virtualize_x18")]
-fn resolve_x18_recovery_pc(
-    resume: X18Resume,
-    site: usize,
-    guest_x18: usize,
-    conditional_target: Option<u64>,
-) -> Option<usize> {
-    Some(match resume {
-        X18Resume::Original => site,
-        X18Resume::Next => site.wrapping_add(4),
-        X18Resume::TakenTarget => conditional_target?.trunc(),
-        X18Resume::LogicalX18 => guest_x18,
-    })
-}
-
 /// Recognizes and canonicalizes an interrupted compact gate using only bounded,
 /// caller-supplied reads. The signal-handler caller supplies exception-table
 /// reads; tests supply synthetic mappings.
@@ -1489,505 +1373,49 @@ pub(super) fn canonicalize_aarch64_gate_signal_context(
     )
 }
 
+fn canonicalize_aarch64_gate_signal_context_with_kind_and_copy(
+    context: &libc::ucontext_t,
+    saved: &litebox_common_linux::PtRegs,
+    runtime: GateRuntimeState,
+    interruption: GateInterruption,
+    read: impl FnMut(usize, &mut [u8]) -> bool,
+) -> (Aarch64GateSignalResult, litebox_common_linux::PtRegs) {
+    let mut registers = saved.clone();
+    copy_signal_context(&mut registers, context);
+    let result = litebox_common_linux::gate_recovery::canonicalize(
+        &registers,
+        runtime,
+        interruption,
+        litebox_syscall_rewriter::TargetHost::Linux,
+        cfg!(feature = "aarch64_virtualize_x18"),
+        read,
+    );
+    (result, registers)
+}
+
+#[cfg(test)]
 fn canonicalize_aarch64_gate_signal_context_with_kind(
     context: &libc::ucontext_t,
     saved: &litebox_common_linux::PtRegs,
     runtime: GateRuntimeState,
     interruption: GateInterruption,
-    mut read: impl FnMut(usize, &mut [u8]) -> bool,
+    read: impl FnMut(usize, &mut [u8]) -> bool,
 ) -> Aarch64GateSignalResult {
-    // Frame and slot sizes come from the rewriter; gate-local instruction
-    // boundaries below follow each emitted template.
-    const SVC_FRAME: usize = SVC_FRAME_BYTES as usize;
-    const MSR_FRAME: usize = MSR_FRAME_BYTES as usize;
-    let pc = context.uc_mcontext.pc;
-    if !pc.is_multiple_of(4) {
-        return Aarch64GateSignalResult::NotGate;
-    }
-    let pc: usize = pc.trunc();
-
-    let aligned = pc & !(GATE_ALIGNMENT - 1);
-    let mut found = None;
-    for candidate_index in 0..GATE_PC_CANDIDATE_COUNT {
-        let Some(slot_start) = aligned.checked_sub(candidate_index * GATE_ALIGNMENT) else {
-            continue;
-        };
-        for slot_size in GATE_SLOT_SIZES {
-            let Some(slot_end) = slot_start.checked_add(slot_size) else {
-                continue;
-            };
-            if pc < slot_start || pc >= slot_end {
-                continue;
-            }
-            let mut metadata_bytes = [0u8; 4];
-            if !read(slot_end - 4, &mut metadata_bytes) {
-                continue;
-            }
-            let Some(metadata) = litebox_syscall_rewriter::aarch64::decode_gate_metadata_word(
-                u32::from_le_bytes(metadata_bytes),
-            ) else {
-                continue;
-            };
-            if metadata.slot_size() != slot_size {
-                continue;
-            }
-            let mut slot = [0u8; SVC_SLOT_BYTES];
-            if !read(slot_start, &mut slot[..slot_size]) {
-                continue;
-            }
-            // Everything up to here is a guess drawn from guest memory: the
-            // metadata word lives at `slot_end - 4`, which the guest can write.
-            // Failing to validate is the evidence that this candidate is not a
-            // gate, so it is discarded rather than treated as a corrupt one --
-            // otherwise a guest could stage a plausible metadata word, fault at
-            // a PC of its choosing and take the runtime down at will. This also
-            // covers a PC in a real gate's padding or metadata: that code never
-            // ran the prologue, so the interrupted registers are the guest's
-            // own and delivering the signal unchanged is correct.
-            let Some(classified) = litebox_syscall_rewriter::aarch64::classify_copied_gate_slot(
-                &slot[..slot_size],
-                slot_start as u64,
-                pc as u64,
-            ) else {
-                continue;
-            };
-            if found.is_some() {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            }
-            found = Some((slot_start, classified));
-        }
-    }
-    let Some((slot_start, gate)) = found else {
-        return Aarch64GateSignalResult::NotGate;
-    };
-
-    // Guest-derived provenance failures are `NotGate`. Once a gate is
-    // authenticated, unreadable state required to restore it is invalid.
-    let site = gate.original_site();
-    if !site.is_multiple_of(4) {
-        return Aarch64GateSignalResult::NotGate;
-    }
-    let site: usize = site.trunc();
-    let mut original = [0u8; 4];
-    if !read(site, &mut original) {
-        return Aarch64GateSignalResult::NotGate;
-    }
-    let original = u32::from_le_bytes(original);
-    if litebox_syscall_rewriter::aarch64::decode_branch_target(original, site as u64)
-        != Some(slot_start as u64)
-    {
-        return Aarch64GateSignalResult::NotGate;
-    }
-
-    let metadata = gate.metadata();
-    let offset = pc - slot_start;
-    if matches!(metadata, GateMetadata::Svc) && offset >= SVC_GATE_BYTES {
-        return if runtime.expected_outbound_stub == slot_start + SVC_GATE_BYTES
-            && runtime.expected_outbound_pc == site + 4
-        {
-            Aarch64GateSignalResult::PreserveSavedContext
-        } else {
-            Aarch64GateSignalResult::NotGate
-        };
-    }
-
-    let mut canonical = saved.clone();
-    copy_signal_context(&mut canonical, context);
-    canonical.pc = site;
-
-    // `copy_signal_context` derived `orig_x0` from the interrupted `regs[0]`,
-    // which for `mrs x0, tpidr_el0` is the *host* anchor. Whatever the arms
-    // below repair, re-derive it afterwards rather than leaving a host address
-    // in a guest-visible field.
-    match metadata {
-        GateMetadata::MrsTpidr { destination, .. } => {
-            let destination = usize::from(destination);
-            let Some(stage) = MrsTpidrGateOffset::from_offset(offset) else {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            };
-            let Some(recovery) = stage.recovery_plan() else {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            };
-            if recovery.runtime_access == RuntimeAccess::Memory
-                && interruption == GateInterruption::Synchronous
-            {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            }
-            match recovery.value {
-                MrsTpidrValueSource::Register => {}
-                MrsTpidrValueSource::Slot => {
-                    let Some(guest_tp) = read_usize(&mut read, runtime.guest_thread_pointer_addr)
-                    else {
-                        return Aarch64GateSignalResult::InvalidRuntimeState;
-                    };
-                    canonical.regs[destination] = guest_tp;
-                }
-            }
-            if recovery.completed {
-                canonical.pc = site + 4;
-            }
-        }
-        GateMetadata::MsrTpidr { source, .. } => {
-            let Some(stage) = MsrTpidrGateOffset::from_offset(offset) else {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            };
-            let Some(recovery) = stage.recovery_plan() else {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            };
-            if recovery.runtime_access == RuntimeAccess::Memory
-                && interruption == GateInterruption::Synchronous
-            {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            }
-            let restored = if recovery.frame == MsrTpidrFrameState::RestoreRegisters {
-                let mut frame = [[0u8; size_of::<usize>()]; 3];
-                if !read(canonical.sp, frame.as_flattened_mut()) {
-                    return Aarch64GateSignalResult::NotGate;
-                }
-                let saved_x16 = usize::from_ne_bytes(frame[0]);
-                let saved_x17 = usize::from_ne_bytes(frame[1]);
-                let captured = usize::from_ne_bytes(frame[2]);
-                let architectural_source = match source {
-                    16 => saved_x16,
-                    17 => saved_x17,
-                    31 => 0,
-                    register => canonical.regs[usize::from(register)],
-                };
-                if captured != architectural_source {
-                    return Aarch64GateSignalResult::NotGate;
-                }
-                Some([saved_x16, saved_x17])
-            } else {
-                // Before commit, the staged value cheaply checks consistency
-                // with the interrupted frame. After LDP at +28, recovery needs
-                // neither the frame nor this check.
-                if recovery.validate_capture {
-                    let Some(captured) = read_usize(
-                        &mut read,
-                        canonical.sp.wrapping_add(usize::from(MSR_FRAME_OFF_VALUE)),
-                    ) else {
-                        return Aarch64GateSignalResult::NotGate;
-                    };
-                    let architectural_source = match source {
-                        31 => 0,
-                        register => canonical.regs[usize::from(register)],
-                    };
-                    if captured != architectural_source {
-                        return Aarch64GateSignalResult::NotGate;
-                    }
-                }
-                None
-            };
-            let guest_sp = match recovery.frame {
-                MsrTpidrFrameState::Absent => canonical.sp,
-                MsrTpidrFrameState::RegistersLive | MsrTpidrFrameState::RestoreRegisters => {
-                    canonical.sp.wrapping_add(MSR_FRAME)
-                }
-            };
-            if let Some([saved_x16, saved_x17]) = restored {
-                canonical.regs[16] = saved_x16;
-                canonical.regs[17] = saved_x17;
-            }
-            canonical.sp = guest_sp;
-            if recovery.completed {
-                canonical.pc = site + 4;
-            }
-        }
-        GateMetadata::Svc => {
-            let Some(stage) = SvcGateOffset::from_offset(offset) else {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            };
-            let Some(plan) = stage.recovery_plan() else {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            };
-            // SVC uses a template-fixed literal; supported MRS/MSR execution
-            // uses the permanent host anchor read by the gate itself.
-            if plan.runtime_access == RuntimeAccess::Memory
-                && interruption == GateInterruption::Synchronous
-            {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            }
-            let saved_x16 = match plan.frame {
-                SvcFrameState::NoFrame | SvcFrameState::FrameAtSpX16Live => None,
-                SvcFrameState::RestoreX16FromFrame => {
-                    let Some(saved_x16) = read_usize(&mut read, canonical.sp) else {
-                        return Aarch64GateSignalResult::NotGate;
-                    };
-                    Some(saved_x16)
-                }
-            };
-            let guest_sp = match plan.frame {
-                SvcFrameState::NoFrame => canonical.sp,
-                SvcFrameState::FrameAtSpX16Live | SvcFrameState::RestoreX16FromFrame => {
-                    canonical.sp.wrapping_add(SVC_FRAME)
-                }
-            };
-            if let Some(saved_x16) = saved_x16 {
-                canonical.regs[16] = saved_x16;
-            }
-            canonical.sp = guest_sp;
-        }
-        #[cfg(feature = "aarch64_virtualize_x18")]
-        GateMetadata::X18 { scratch } => {
-            let Some(stage) = X18GateOffset::from_offset(offset) else {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            };
-            let Some(plan) = stage.recovery_plan() else {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            };
-            if plan.runtime_access == RuntimeAccess::Memory
-                && interruption == GateInterruption::Synchronous
-            {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            }
-            let Some((guest_sp, restored_registers)) =
-                recover_x18_frame(plan.frame, canonical.sp, &mut read)
-            else {
-                return Aarch64GateSignalResult::NotGate;
-            };
-            let guest_x18 = match plan.value {
-                X18ValueSource::Scratch => context.uc_mcontext.regs[usize::from(scratch)].trunc(),
-                X18ValueSource::Slot => {
-                    let Some(value) = read_usize(
-                        &mut read,
-                        runtime.guest_thread_pointer_addr
-                            + litebox_syscall_rewriter::aarch64::GUEST_X18_OFFSET_FROM_GUEST_TP,
-                    ) else {
-                        return Aarch64GateSignalResult::InvalidRuntimeState;
-                    };
-                    value
-                }
-            };
-            let Some(resume_pc) = resolve_x18_recovery_pc(plan.resume, site, guest_x18, None)
-            else {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            };
-            apply_x18_recovery(
-                &mut canonical,
-                scratch,
-                gate.anchor_scratch().unwrap(),
-                restored_registers,
-                guest_sp,
-                guest_x18,
-                resume_pc,
-            );
-        }
-        #[cfg(feature = "aarch64_virtualize_x18")]
-        GateMetadata::X18StackWriteback { scratch } => {
-            let Some(plan) = X18StackWritebackOffset::from_offset(offset)
-                .and_then(X18StackWritebackOffset::recovery_plan)
-            else {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            };
-            let Some((delta, frame_bytes)) = gate.stack_writeback() else {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            };
-            if interruption == GateInterruption::Synchronous
-                && plan.fault_attribution != X18FaultAttribution::GuestInstruction
-            {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            }
-
-            let anchor = gate.anchor_scratch().unwrap();
-            let frame_bytes = usize::from(frame_bytes);
-            let frame = match plan.frame {
-                X18StackWritebackFrameState::Absent => None,
-                X18StackWritebackFrameState::AtSpRegistersLive
-                | X18StackWritebackFrameState::AtSpRestoreRegisters
-                | X18StackWritebackFrameState::AtSpRegistersRestored => Some(canonical.sp),
-                X18StackWritebackFrameState::AtAnchorRestoreRegisters => {
-                    Some(context.uc_mcontext.regs[usize::from(anchor)].trunc())
-                }
-            };
-            let restored = match plan.frame {
-                X18StackWritebackFrameState::AtSpRestoreRegisters
-                | X18StackWritebackFrameState::AtAnchorRestoreRegisters => {
-                    const SCRATCH_REGISTER_COUNT: usize = 2;
-                    let mut words = [[0u8; size_of::<usize>()]; SCRATCH_REGISTER_COUNT];
-                    if !read(frame.unwrap(), words.as_flattened_mut()) {
-                        return Aarch64GateSignalResult::InvalidRuntimeState;
-                    }
-                    Some(words.map(usize::from_ne_bytes))
-                }
-                X18StackWritebackFrameState::Absent
-                | X18StackWritebackFrameState::AtSpRegistersLive
-                | X18StackWritebackFrameState::AtSpRegistersRestored => None,
-            };
-            let guest_sp = match plan.sp {
-                X18StackWritebackSpSource::Signal => canonical.sp,
-                X18StackWritebackSpSource::FramePlusOriginal => {
-                    frame.unwrap().wrapping_add(frame_bytes)
-                }
-                X18StackWritebackSpSource::FramePlusResult => frame
-                    .unwrap()
-                    .wrapping_add(frame_bytes)
-                    .wrapping_add_signed(isize::from(delta)),
-            };
-            let guest_x18 = match plan.value {
-                X18ValueSource::Scratch => {
-                    // At the transformed instruction this is the exact value
-                    // reported by hardware; no completion inference is made.
-                    context.uc_mcontext.regs[usize::from(scratch)].trunc()
-                }
-                X18ValueSource::Slot => {
-                    let Some(value) = read_usize(
-                        &mut read,
-                        runtime.guest_thread_pointer_addr
-                            + litebox_syscall_rewriter::aarch64::GUEST_X18_OFFSET_FROM_GUEST_TP,
-                    ) else {
-                        return Aarch64GateSignalResult::InvalidRuntimeState;
-                    };
-                    value
-                }
-            };
-            let Some(resume_pc) = resolve_x18_recovery_pc(plan.resume, site, guest_x18, None)
-            else {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            };
-            apply_x18_recovery(
-                &mut canonical,
-                scratch,
-                anchor,
-                restored,
-                guest_sp,
-                guest_x18,
-                resume_pc,
-            );
-        }
-        #[cfg(feature = "aarch64_virtualize_x18")]
-        GateMetadata::X18CompareBranch { scratch } => {
-            let Some(stage) = X18CompareBranchOffset::from_offset(offset) else {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            };
-            let Some(plan) = stage.recovery_plan() else {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            };
-            if plan.runtime_access == RuntimeAccess::Memory
-                && interruption == GateInterruption::Synchronous
-            {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            }
-            let Some((guest_sp, restored_registers)) =
-                recover_x18_frame(plan.frame, canonical.sp, &mut read)
-            else {
-                return Aarch64GateSignalResult::NotGate;
-            };
-            let Some(guest_x18) = read_usize(
-                &mut read,
-                runtime.guest_thread_pointer_addr
-                    + litebox_syscall_rewriter::aarch64::GUEST_X18_OFFSET_FROM_GUEST_TP,
-            ) else {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            };
-            let Some(resume_pc) =
-                resolve_x18_recovery_pc(plan.resume, site, guest_x18, gate.conditional_target())
-            else {
-                return Aarch64GateSignalResult::NotGate;
-            };
-            apply_x18_recovery(
-                &mut canonical,
-                scratch,
-                gate.anchor_scratch().unwrap(),
-                restored_registers,
-                guest_sp,
-                guest_x18,
-                resume_pc,
-            );
-        }
-        #[cfg(feature = "aarch64_virtualize_x18")]
-        GateMetadata::X18Adr { scratch } => {
-            let Some(stage) = X18AdrOffset::from_offset(offset) else {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            };
-            let Some(plan) = stage.recovery_plan() else {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            };
-            if plan.runtime_access == RuntimeAccess::Memory
-                && interruption == GateInterruption::Synchronous
-            {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            }
-            let Some((guest_sp, restored_registers)) =
-                recover_x18_frame(plan.frame, canonical.sp, &mut read)
-            else {
-                return Aarch64GateSignalResult::NotGate;
-            };
-            let guest_x18 = match plan.value {
-                X18ValueSource::Scratch => context.uc_mcontext.regs[usize::from(scratch)].trunc(),
-                X18ValueSource::Slot => {
-                    let Some(value) = read_usize(
-                        &mut read,
-                        runtime.guest_thread_pointer_addr
-                            + litebox_syscall_rewriter::aarch64::GUEST_X18_OFFSET_FROM_GUEST_TP,
-                    ) else {
-                        return Aarch64GateSignalResult::InvalidRuntimeState;
-                    };
-                    value
-                }
-            };
-            let Some(resume_pc) = resolve_x18_recovery_pc(plan.resume, site, guest_x18, None)
-            else {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            };
-            apply_x18_recovery(
-                &mut canonical,
-                scratch,
-                gate.anchor_scratch().unwrap(),
-                restored_registers,
-                guest_sp,
-                guest_x18,
-                resume_pc,
-            );
-        }
-        #[cfg(feature = "aarch64_virtualize_x18")]
-        GateMetadata::X18Branch { kind } => {
-            let Some(plan) = x18_branch_recovery_plan(offset) else {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            };
-            if plan.frame != X18FrameState::Absent
-                || plan.value != X18ValueSource::Slot
-                || plan.resume != X18Resume::LogicalX18
-                || plan.runtime_access != RuntimeAccess::NoAccess
-            {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            }
-            let Some(guest_x18) = read_usize(
-                &mut read,
-                runtime.guest_thread_pointer_addr
-                    + litebox_syscall_rewriter::aarch64::GUEST_X18_OFFSET_FROM_GUEST_TP,
-            ) else {
-                return Aarch64GateSignalResult::InvalidRuntimeState;
-            };
-            canonical.regs[18] = guest_x18;
-            if interruption == GateInterruption::Breakpoint {
-                if kind.writes_link_register() {
-                    canonical.regs[30] = site.wrapping_add(4);
-                }
-                let Some(resume_pc) = resolve_x18_recovery_pc(plan.resume, site, guest_x18, None)
-                else {
-                    return Aarch64GateSignalResult::InvalidRuntimeState;
-                };
-                canonical.pc = resume_pc;
-                canonical.orig_x0 = canonical.regs[0];
-                return Aarch64GateSignalResult::ResumeGuest(canonical);
-            }
-        }
-        #[cfg(not(feature = "aarch64_virtualize_x18"))]
-        GateMetadata::X18 { .. }
-        | GateMetadata::X18StackWriteback { .. }
-        | GateMetadata::X18CompareBranch { .. }
-        | GateMetadata::X18Adr { .. }
-        | GateMetadata::X18Branch { .. } => return Aarch64GateSignalResult::NotGate,
-    }
-    canonical.orig_x0 = canonical.regs[0];
-    Aarch64GateSignalResult::Canonicalized(canonical)
+    canonicalize_aarch64_gate_signal_context_with_kind_and_copy(
+        context,
+        saved,
+        runtime,
+        interruption,
+        read,
+    )
+    .0
 }
 
 pub(super) fn canonicalize_runtime_aarch64_gate_signal_context(
     context: &libc::ucontext_t,
     saved: &litebox_common_linux::PtRegs,
     interruption: GateInterruption,
-) -> Aarch64GateSignalResult {
+) -> (Aarch64GateSignalResult, litebox_common_linux::PtRegs) {
     let block: usize;
     let expected_outbound_stub: usize;
     let expected_outbound_pc: usize;
@@ -2009,7 +1437,7 @@ pub(super) fn canonicalize_runtime_aarch64_gate_signal_context(
             options(nostack, readonly, preserves_flags)
         );
     }
-    canonicalize_aarch64_gate_signal_context_with_kind(
+    canonicalize_aarch64_gate_signal_context_with_kind_and_copy(
         context,
         saved,
         GateRuntimeState {
